Truncate fails on mysql tables that have foreign key constraints.  This is the recommended solution.

+++ lib/database_cleaner/sequel/truncation.rb
@@ -9,7 +9,7 @@ module DatabaseCleaner
   
       def clean
         case db.database_type
-        when :postgres
+        when :postgres;
           # PostgreSQL requires all tables with FKs to be truncates in the same command, or have the CASCADE keyword
           # appended. Bulk truncation without CASCADE is:
           # * Safer. Tables outside of tables_to_truncate won't be affected.
@@ -18,6 +18,14 @@ module DatabaseCleaner
             all_tables= tables.map{|t| %["#{t}"]}.join ','
             db.run "TRUNCATE TABLE #{all_tables};"
           end
+        when :mysql;
+          # Make sure to turn off foreign key checks, or truncate command fails
+          db.run "set FOREIGN_KEY_CHECKS=0;"
+          # Truncate each table normally
+          each_table do |db, table|
+            db[table].truncate
+          end
+          db.run "set FOREIGN_KEY_CHECKS=1;"
         else
           # Truncate each table normally
           each_table do |db, table|
